requirements.txt
bcrypt==4.0.1
blinker==1.6.2
bpyutils==0.5.8
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
mysqlclient==2.1.1
protobuf==4.23.0
PyJWT==2.7.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
Werkzeug==2.3.4
